[#134] Fix error when submitting board
When submitting an order, the app tries to serialize the board into a
JSON object to be stored in the database. However, we updated the Strip
model at some point to keep track of its Three.js Mesh which can't be
serialized. This was causing the app to error out when submitting. Since
we don't need the Mesh to be stored in the database, we were able to
simply delete the Mesh from each Strip when serializing.